fix: margin trading estimations
